documentation/requirements.txt
djangorestframework==3.16.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
